chore(ingestion-api): decompose monolithic bootstrap into DI modules
Sprint A - Runtime Bootstrap Decomposition:
- bootstrap/create-runtime-context.ts  : Stateful singleton DI root
- bootstrap/create-websocket-gateway.ts: Isolated WS factory
- bootstrap/register-event-observers.ts: Idempotency-guarded firehose observer
- bootstrap/create-express-app.ts      : Express + all routes, deps injected
- bootstrap/create-http-server.ts      : Port binding + EADDRINUSE guard
- index.ts                             : Pure 6-step DI orchestration (391 -> 68 lines)
- register-command-handlers.ts         : sseManager singleton removed, now injected

Sprint B - Production Env Guard:
- security/origin-policy.ts            : Hard boot fail if ALLOWED_ORIGINS missing in prod
- scripts/active/audit-production-env.mjs: CI env contract audit script

Typecheck: 26 errors -> 21 (5 bootstrap errors fixed, 21 pre-existing deferred)
Build: admin-panel built in 4.63s - no regression

Pre-existing type drift deferred to Sprint C:
  oracle-cognitive-decision.contract.js missing
  telemetry.event_captured union gap
  boardroomReducer export missing
  replay-engine @santis/db imports broken
  bus.publish API drift

Closes #135
